vertex/ -> vertex.hpp
